successfully implement modules
